+++ src/cp_abe/ac17.rs
@@ -160,7 +160,7 @@ mod test {
             let result = rabe_cp_ac17_decrypt(cipher, secret_key);
             assert!(!result.buffer.is_null());
             assert_eq!(
-                std::slice::from_raw_parts(result.buffer, result.len),
+                std::slice::from_raw_parts(result.buffer, result.len as usize),
                 "hello world".as_bytes()
             );
